Add zod-bird and recordLink
