+++ apps/web/app/api/cron/domains/transfer/utils.ts
@@ -1,5 +1,4 @@
 import prisma from "@/lib/prisma";
-import { recordLink } from "@/lib/tinybird";
 import { formatRedisLink, redis } from "@/lib/upstash";
 import { Link } from "@prisma/client";
 import { sendEmail } from "emails";
@@ -92,21 +91,4 @@ export const domainTransferredEmail = async ({
       linksCount,
     }),
   });
-};
-
-// Record link in Tinybird
-export const recordLinks = async ({
-  newWorkspaceId,
-  links,
-}: {
-  newWorkspaceId: string;
-  links: Link[];
-}) => {
-  await Promise.all(
-    links.map((link) =>
-      recordLink({ link: { ...link, projectId: newWorkspaceId } }),
-    ),
-  );
-
-  return;
 };
