+++ apps/web/app/api/links/[linkId]/transfer/route.ts
@@ -95,10 +95,12 @@ export const POST = withWorkspace(
           }),
         }),
         recordLink({
-          link: {
-            ...link,
-            projectId: newWorkspaceId,
-          },
+          link_id: link.id,
+          domain: link.domain,
+          key: link.key,
+          url: link.url,
+          tagIds: [],
+          project_id: newWorkspaceId,
         }),
         // decrement old workspace usage
         prisma.project.update({
