+++ apps/web/app/api/tags/[id]/route.ts
@@ -59,7 +59,18 @@ export const DELETE = withWorkspace(async ({ params, workspace }) => {
         projectId: workspace.id,
       },
       include: {
-        links: true,
+        links: {
+          select: {
+            link: {
+              select: {
+                id: true,
+                domain: true,
+                key: true,
+                url: true,
+              },
+            },
+          },
+        },
       },
     });
 
@@ -71,21 +82,15 @@ export const DELETE = withWorkspace(async ({ params, workspace }) => {
     }
 
     // update links metadata in tinybird after deleting a tag
-    await Promise.all(
-      response.links.map(async ({ linkId }) => {
-        const link = await prisma.link.findUnique({
-          where: {
-            id: linkId,
-          },
-          include: {
-            tags: true,
-          },
-        });
-        if (!link) {
-          return null;
-        }
-        return await recordLink({ link });
-      }),
+    await recordLink(
+      response.links.map(({ link }) => ({
+        link_id: link.id,
+        domain: link.domain,
+        key: link.key,
+        url: link.url,
+        tagIds: [],
+        project_id: workspace.id,
+      })),
     );
 
     return NextResponse.json(response);
