+++ apps/web/lib/api/domains/get-config-response.ts
@@ -1,0 +1,12 @@
+export const getConfigResponse = async (domain: string) => {
+  return await fetch(
+    `https://api.vercel.com/v6/domains/${domain.toLowerCase()}/config?teamId=${process.env.TEAM_ID_VERCEL}`,
+    {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${process.env.AUTH_BEARER_TOKEN}`,
+        "Content-Type": "application/json",
+      },
+    },
+  ).then((res) => res.json());
+};
