+++ apps/web/lib/api/domains/get-domain-response.ts
@@ -1,0 +1,14 @@
+export const getDomainResponse = async (domain: string) => {
+  return await fetch(
+    `https://api.vercel.com/v9/projects/${process.env.PROJECT_ID_VERCEL}/domains/${domain.toLowerCase()}?teamId=${process.env.TEAM_ID_VERCEL}`,
+    {
+      method: "GET",
+      headers: {
+        Authorization: `Bearer ${process.env.AUTH_BEARER_TOKEN}`,
+        "Content-Type": "application/json",
+      },
+    },
+  ).then((res) => {
+    return res.json();
+  });
+};
