+++ apps/web/lib/api/domains/index.ts
@@ -1,0 +1,8 @@
+export * from "./add-domain-vercel";
+export * from "./delete-domain-links";
+export * from "./get-config-response";
+export * from "./get-domain-response";
+export * from "./remove-domain-vercel";
+export * from "./set-root-domain";
+export * from "./utils";
+export * from "./verify-domain";
