+++ apps/web/lib/api/domains/remove-domain-vercel.ts
@@ -1,0 +1,45 @@
+import prisma from "@/lib/prisma";
+import { getApexDomain } from "@dub/utils";
+
+export const removeDomainFromVercel = async (domain: string) => {
+  const apexDomain = getApexDomain(`https://${domain}`);
+  const domainCount = await prisma.domain.count({
+    where: {
+      OR: [
+        {
+          slug: apexDomain,
+        },
+        {
+          slug: {
+            endsWith: `.${apexDomain}`,
+          },
+        },
+      ],
+    },
+  });
+  if (domainCount > 1) {
+    // the apex domain is being used by other domains on Dub
+    // so we should only remove it from our Vercel project
+    return await fetch(
+      `https://api.vercel.com/v9/projects/${process.env.PROJECT_ID_VERCEL}/domains/${domain.toLowerCase()}?teamId=${process.env.TEAM_ID_VERCEL}`,
+      {
+        headers: {
+          Authorization: `Bearer ${process.env.AUTH_BEARER_TOKEN}`,
+        },
+        method: "DELETE",
+      },
+    ).then((res) => res.json());
+  } else {
+    // this is the only domain using this apex domain
+    // so we can remove it entirely from our Vercel team
+    return await fetch(
+      `https://api.vercel.com/v6/domains/${domain.toLowerCase()}?teamId=${process.env.TEAM_ID_VERCEL}`,
+      {
+        headers: {
+          Authorization: `Bearer ${process.env.AUTH_BEARER_TOKEN}`,
+        },
+        method: "DELETE",
+      },
+    ).then((res) => res.json());
+  }
+};
