+++ apps/web/lib/api/domains/verify-domain.ts
@@ -1,0 +1,12 @@
+export const verifyDomain = async (domain: string) => {
+  return await fetch(
+    `https://api.vercel.com/v9/projects/${process.env.PROJECT_ID_VERCEL}/domains/${domain.toLowerCase()}/verify?teamId=${process.env.TEAM_ID_VERCEL}`,
+    {
+      method: "POST",
+      headers: {
+        Authorization: `Bearer ${process.env.AUTH_BEARER_TOKEN}`,
+        "Content-Type": "application/json",
+      },
+    },
+  ).then((res) => res.json());
+};
