+++ apps/web/lib/api/links/delete-link.ts
@@ -20,7 +20,15 @@ export async function deleteLink(linkId: string) {
         link.image?.startsWith(process.env.STORAGE_BASE_URL as string) &&
         storage.delete(`images/${link.id}`),
       redis.hdel(link.domain.toLowerCase(), link.key.toLowerCase()),
-      recordLink({ link, deleted: true }),
+      recordLink({
+        link_id: link.id,
+        domain: link.domain,
+        key: link.key,
+        url: link.url,
+        tagIds: link.tags.map((tag) => tag.tagId),
+        project_id: link.projectId,
+        deleted: true,
+      }),
       link.projectId &&
         prisma.project.update({
           where: {
