+++ apps/web/lib/tinybird/client.ts
@@ -1,0 +1,6 @@
+import { Tinybird } from "@chronark/zod-bird";
+
+export const tb = new Tinybird({
+  token: process.env.TINYBIRD_API_KEY as string,
+  baseUrl: process.env.TINYBIRD_API_URL as string,
+});
