+++ apps/web/lib/tinybird/index.ts
@@ -1,2 +1,3 @@
+export * from "./client";
 export * from "./record-click";
 export * from "./record-link";
