+++ apps/web/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@anthropic-ai/sdk": "^0.20.4",
     "@boxyhq/saml-jackson": "1.23.6",
+    "@chronark/zod-bird": "^0.3.9",
     "@dub/tailwind-config": "workspace:*",
     "@dub/ui": "workspace:*",
     "@dub/utils": "workspace:*",
